requirements.txt
python-dotenv
requests
simple-salesforce
